reformage de code
ajout de pathlib, déplacement de Player dans son fichier, compréhension de listes
