+++ src/utils.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+ROOT_DIR = Path(__file__).parent.parent
+IMAGES_DIR = ROOT_DIR.joinpath("images")
+ENIGMA_DIR = ROOT_DIR.joinpath("enigmes")
